Improves project structure
