+++ .github/workflows/publish-bundles.yaml
@@ -23,12 +23,10 @@ jobs:
       - name: install requirements
         run: pip install -r requirements.txt
       - name: render bundle.yaml
-        run: |
-          python src/render_bundle.py --output_file bundle.yaml --bundle_variant=${{ inputs.bundle_variant }}
-        env:
-          PYTHONPATH: $(pwd)
+        run: python3 render_bundle/render_bundle.py --output_file bundle/bundle.yaml --bundle_variant=${{ inputs.bundle_variant }}
       - name: Upload bundle
         uses: canonical/charming-actions/upload-bundle@2.2.5
         with:
           credentials: "${{ secrets.CHARMCRAFT_AUTH }}"
+          bundle-path: "bundle"
           github-token: "${{ secrets.GITHUB_TOKEN }}"
