--- charmcraft.yaml
@@ -1,1 +1,0 @@
-type: bundle
