+++ render_bundle/lib/charm_bundle_generator.py
@@ -3,14 +3,15 @@
 
 """Generates any charm bundle from user provided Application and Relation objects."""
 
-from pathlib import Path
 from typing import Optional
 
 from jinja2 import Template
 from pydantic import BaseModel
+import os
+import sys
 
 
-def _write_to_file(file: str, content: str):
+def write_to_file(file: str, content: str):
     """Writes content to a file.
 
     Args:
@@ -65,28 +66,19 @@ class CharmBundle:
         self.applications = applications
         self.relations = relations
 
-    def render(self) -> str:
+    def render(self, output_file: str):
         """Renders charm bundle using jinja2 templating.
 
-        Returns:
-            str: Rendered bundle
+        Args:
+            output_file: Output file
         """
-        p = Path(__file__).with_name("bundle.yaml.j2")
-        with p.open("r") as f:
-            template = Template(f.read())
-        return template.render(
+        with open(os.path.join(sys.path[0], "lib/bundle.yaml.j2")) as file:
+            template = Template(file.read())
+        content = template.render(
             bundle=self.bundle,
             name=self.name,
             description=self.description,
             applications=self.applications,
             relations=self.relations,
         )
-
-    def render_to_file(self, output_file: str):
-        """Renders charm bundle using jinja2 templating and outputs to a file.
-
-        Args:
-            output_file: Output file.
-        """
-        content = self.render()
-        _write_to_file(content=content, file=output_file)
+        write_to_file(content=content, file=output_file)
