+++ render_bundle/render_bundle.py
@@ -1,30 +1,28 @@
 #!/usr/bin/env python3
-# Copyright 2023 Canonical Ltd.
+# Copyright 2021 Canonical Ltd.
 # See LICENSE file for licensing details.
 
 
-"""Python script to generate one of the potential SD-Core bundle variants."""
+"""Python script to generate one of the potential SD-CORE bundle variant."""
 
 import argparse
 import enum
-from typing import Optional
 
-from bundles import SDCoreUserPlane
+from bundles import SDCoreUP
 
 
 class SDCoreBundleVariant(enum.Enum):
     """Possible bundles to generate."""
 
-    SDCORE_USER_PLANE = SDCoreUserPlane
+    SDCORE_UP = SDCoreUP
 
 
-def _parse_args() -> tuple[bool, str, str, str]:
+def _parse_args() -> tuple[bool, str, str]:
     """Parses user provided arguments.
 
     Returns:
-        local (bool): Whether to generate a bundle for local charms (instead of charmhub)
-        bundle_variant (str): SD-Core bundle to generate
-        channel(str): Charmhub channel. Only use when "local" is set to False.
+        local (str): Whether to generate a bundle for local charms (instead of charmhub)
+        bundle_variant (str): SD-CORE bundle to generate
         output_file (str): Directory and file name where bundle will be generated.
     """
     parser = argparse.ArgumentParser(description="Render jinja2 bundle template from cli args.")
@@ -40,15 +38,7 @@ def _parse_args() -> tuple[bool, str, str, str]:
         type=str,
         help="SDCORE bundle type.",
         required=True,
-        choices=["SDCORE_USER_PLANE"],
-    )
-    parser.add_argument(
-        "--channel",
-        type=str,
-        help="Output file",
-        choices=["edge", "beta", "candidate", "stable"],
-        required=False,
-        default="edge",
+        choices=["SDCORE_UP"],
     )
     parser.add_argument(
         "--output_file",
@@ -58,42 +48,20 @@ def _parse_args() -> tuple[bool, str, str, str]:
         default="bundle.yaml",
     )
     bundle_args, _ = parser.parse_known_args()
-    return (
-        bundle_args.local,
-        bundle_args.bundle_variant,
-        bundle_args.channel,
-        bundle_args.output_file,
-    )
+    return bundle_args.local, bundle_args.bundle_variant, bundle_args.output_file
 
 
-def render_bundle(
-    local: bool,
-    bundle_variant: str,
-    output_file: str,
-    channel: Optional[str] = None,
-):
-    """Generates a SD-Core bundle variant.
-
-    Args:
-        local: Whether to use local charms (instead of charmhub).
-        bundle_variant: Bundle variant (ex. SDCORE_USER_PLANE)
-        channel: Charmhub channel
-        output_file: Output file path and directory (ex. `my/path/bundle.yaml`)
-    """
+def render_bundle(local: bool, bundle_variant: str, output_file: str):
+    """Generates bundle."""
     bundle_variant_type = SDCoreBundleVariant[bundle_variant]
-    bundle = bundle_variant_type.value(local=local, channel=channel)
-    bundle.render_to_file(output_file=output_file)
+    bundle = bundle_variant_type.value(local=local)
+    bundle.render(output_file=output_file)
 
 
 def main():
     """Generates one of the SDCORE charm bundles based on user provided bundle type."""
-    local, bundle_variant, channel, output_file = _parse_args()
-    render_bundle(
-        local=local,
-        bundle_variant=bundle_variant,
-        channel=channel,
-        output_file=output_file,
-    )
+    local, bundle_variant, output_file = _parse_args()
+    render_bundle(local=local, bundle_variant=bundle_variant, output_file=output_file)
 
 
 if __name__ == "__main__":
