+++ render_bundle/tests/unit/lib/expected_bundle_charmhub.yaml
@@ -5,13 +5,15 @@ applications:
   app-1-name:
     charm: app-1-charm
     scale: 1
+    trust: False
   app-2-name:
     charm: app-2-charm
     scale: 1
-    trust: true
+    trust: False
   app-3-name:
     charm: app-3-charm
     scale: 1
+    trust: False
 relations:
   - - app-1-name:banana
     - app-2-name:fruit
