+++ render_bundle/tests/unit/src/expected_bundle_sdcore_up_local.yaml
