+++ render_bundle/tox.ini
@@ -4,16 +4,17 @@
 [tox]
 skipsdist=True
 skip_missing_interpreters = True
-envlist = fmt, lint, unit, static
+envlist = lint, unit, static
 
 [vars]
 lib_path = {toxinidir}/lib/
-src_path = {toxinidir}
-unit_tests_path = {toxinidir}/tests/unit/
-all_path = {[vars]src_path} {[vars]lib_path} {[vars]unit_tests_path}
+src_path = {toxinidir}/src/
+unit_test_path = {toxinidir}/tests/unit/
+all_path = {[vars]src_path} {[vars]lib_path} {[vars]unit_test_path}
 
 [testenv]
 setenv =
+  PYTHONPATH = {toxinidir}:{[vars]lib_path}:{[vars]src_path}
   PYTHONBREAKPOINT=pdb.set_trace
   PY_COLORS=1
 deps =
@@ -27,23 +28,23 @@ passenv =
 [testenv:fmt]
 description = Apply coding style standards to code
 commands =
-    isort {[vars]src_path}
-    black {[vars]src_path}
+    isort {[vars]all_path}
+    black {[vars]all_path}
 
 [testenv:lint]
 description = Check code against coding style standards
 commands =
-    pflake8 {[vars]src_path}
-    isort --check-only --diff {[vars]src_path}
-    black --check --diff {[vars]src_path}
+    pflake8 {[vars]all_path}
+    isort --check-only --diff {[vars]all_path}
+    black --check --diff {[vars]all_path}
 
 [testenv:static]
 description = Run static analysis checks
 commands =
-    mypy {[vars]src_path} {posargs}
+    mypy src/ {posargs}
 
 [testenv:unit]
 description = Run unit tests
 commands =
-    coverage run --source={[vars]src_path} --omit={[vars]unit_tests_path}/* -m pytest {[vars]unit_tests_path} -v --tb native -s {posargs}
+    coverage run --source={[vars]src_path} -m pytest {[vars]unit_test_path} -v --tb native -s {posargs}
     coverage report
