--- test-requirements.txt
@@ -1,13 +1,0 @@
-black
-coverage[toml]
-flake8-docstrings
-flake8-builtins
-isort
-mypy
-pep8-naming
-pyproject-flake8
-pytest
-pytest-operator
-types-PyYAML
-types-setuptools
-types-toml
