--- tox.ini
@@ -1,50 +1,0 @@
-# Copyright 2023 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-[tox]
-skipsdist=True
-skip_missing_interpreters = True
-envlist = lint, unit, static
-
-[vars]
-lib_path = {toxinidir}/lib/
-src_path = {toxinidir}/src/
-unit_test_path = {toxinidir}/tests/unit/
-all_path = {[vars]src_path} {[vars]lib_path} {[vars]unit_test_path}
-
-[testenv]
-setenv =
-  PYTHONPATH = {toxinidir}:{[vars]lib_path}:{[vars]src_path}
-  PYTHONBREAKPOINT=pdb.set_trace
-  PY_COLORS=1
-deps =
-    -r{toxinidir}/requirements.txt
-    -r{toxinidir}/test-requirements.txt
-passenv =
-  PYTHONPATH
-  CHARM_BUILD_DIR
-  MODEL_SETTINGS
-
-[testenv:fmt]
-description = Apply coding style standards to code
-commands =
-    isort {[vars]all_path}
-    black {[vars]all_path}
-
-[testenv:lint]
-description = Check code against coding style standards
-commands =
-    pflake8 {[vars]all_path}
-    isort --check-only --diff {[vars]all_path}
-    black --check --diff {[vars]all_path}
-
-[testenv:static]
-description = Run static analysis checks
-commands =
-    mypy src/ {posargs}
-
-[testenv:unit]
-description = Run unit tests
-commands =
-    coverage run --source={[vars]src_path} -m pytest {[vars]unit_test_path} -v --tb native -s {posargs}
-    coverage report
